Sets the FileDialog and SaveDialog Window Locations
FileDialog's position is based on the last open window's top left corner.

Without using DLL imports, the easiest way I found to do this is to create an invisible window before opening the FileDialog.

This solution isn't clean or ideal but it does position the filedialog window without the user knowing what is going on behind the scenes.